+++ firmware/src/modes/GameOfLifeMode.cpp
@@ -46,7 +46,8 @@ void GameOfLifeMode::handle()
             pending = false;
         }
         std::vector<bool> seeds(GRID_COLUMNS * (GRID_ROWS - (clock ? 5 : 0)), false);
-        for (uint8_t i = active; i < GRID_COLUMNS * (GRID_ROWS - (clock ? 5 : 0)) / (1U << 4U); ++i)
+        for (uint8_t i = active; i < static_cast<uint8_t>(GRID_COLUMNS * (GRID_ROWS - (clock ? 5 : 0)) / (1U << 4U));
+             ++i)
         {
             seeds[random(1, GRID_COLUMNS - 1) +
                   (random(clock ? 6 : 1, GRID_ROWS - 1) * (GRID_COLUMNS - (clock ? 5 : 0)))] = true;
@@ -96,7 +97,7 @@ void GameOfLifeMode::setClock(bool _clock)
     nvs_handle_t handle{};
     if (nvs_open(std::string(name).c_str(), nvs_open_mode_t::NVS_READWRITE, &handle) == ESP_OK)
     {
-        nvs_set_u8(handle, "clock", static_cast<uint8_t>(clock));
+        nvs_set_u8(handle, "clock", static_cast<uint8_t>(clock)); // NOLINT(readability-implicit-bool-conversion)
         nvs_commit(handle);
         nvs_close(handle);
     }
